fix: panic on bot_takeover events

+++ pkg/demoinfocs/common/hostage.go
@@ -57,7 +57,7 @@ func (hostage *Hostage) Leader() *Player {
 		return hostage.demoInfoProvider.FindPlayerByPawnHandle(getUInt64(hostage.Entity, "m_leader"))
 	}
 
-	return hostage.demoInfoProvider.FindPlayerByHandle(getInt(hostage.Entity, "m_leader"))
+	return hostage.demoInfoProvider.FindPlayerByHandle(uint64(getInt(hostage.Entity, "m_leader")))
 }
 
 // NewHostage creates a hostage.
